+++ tests/mock_db/mvalue.rs
@@ -27,40 +27,28 @@ impl fmt::Display for Timestamp {
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum MValue {
+    Null,
     Short(i16),
-    NullableShort(Option<i16>),
     Double(f64),
-    NullableDouble(Option<f64>),
     String(String),
-    NullableString(Option<String>),
     Timestamp(Timestamp),
-    NullableTimestamp(Option<Timestamp>),
 }
 
 impl MValue {
+    pub fn new_null() -> MValue {
+        MValue::Null
+    }
     pub fn new_short(i: i16) -> MValue {
         MValue::Short(i)
-    }
-    pub fn new_nullable_short(o_i: Option<i16>) -> MValue {
-        MValue::NullableShort(o_i)
     }
     pub fn new_double(f: f64) -> MValue {
         MValue::Double(f)
     }
-    pub fn new_nullable_double(o_f: Option<f64>) -> MValue {
-        MValue::NullableDouble(o_f)
-    }
     pub fn new_string(s: String) -> MValue {
         MValue::String(s)
     }
-    pub fn new_nullable_string(s: String) -> MValue {
-        MValue::NullableString(Some(s))
-    }
     pub fn new_ts(ts: NaiveDateTime) -> MValue {
         MValue::Timestamp(Timestamp(ts))
-    }
-    pub fn new_nullable_ts(ts: NaiveDateTime) -> MValue {
-        MValue::NullableTimestamp(Some(Timestamp(ts)))
     }
 
     pub fn try_into<'de, T>(self) -> mock_db::Result<T>