+++ tests/mock_db/serde_db_impl/de.rs
@@ -11,10 +11,10 @@ fn not_implemented(s: &'static str) -> ConversionError {
 
 impl DbValue for MValue {
     fn is_null(&self) -> bool {
-        trace!("is_null() for {:?}", self);
-        match *self {
-            MValue::NullableDouble(None) | MValue::NullableShort(None) => true,
-            _ => false,
+        if let MValue::Null = *self {
+            true
+        } else {
+            false
         }
     }
 }
@@ -52,7 +52,7 @@ impl DbValueInto<i8> for MValue {
 impl DbValueInto<i16> for MValue {
     fn try_into(self) -> Result<i16, ConversionError> {
         match self {
-            MValue::Short(i) | MValue::NullableShort(Some(i)) => Ok(i),
+            MValue::Short(i) => Ok(i),
             mv => Err(ConversionError::ValueType(format!(
                 "DbValueInto<i16> not implemented for {:?}",
                 mv
@@ -63,7 +63,7 @@ impl DbValueInto<i16> for MValue {
 impl DbValueInto<i32> for MValue {
     fn try_into(self) -> Result<i32, ConversionError> {
         match self {
-            MValue::Short(i) | MValue::NullableShort(Some(i)) => Ok(i32::from(i)),
+            MValue::Short(i) => Ok(i32::from(i)),
             mv => Err(ConversionError::ValueType(format!(
                 "DbValueInto<i32> not implemented for {:?}",
                 mv
@@ -84,7 +84,7 @@ impl DbValueInto<f32> for MValue {
 impl DbValueInto<f64> for MValue {
     fn try_into(self) -> Result<f64, ConversionError> {
         match self {
-            MValue::Double(f) | MValue::NullableDouble(Some(f)) => Ok(f),
+            MValue::Double(f) => Ok(f),
             mv => Err(ConversionError::ValueType(format!(
                 "DbValueInto<f64> not implemented for {:?}",
                 mv
@@ -98,7 +98,7 @@ impl DbValueInto<String> for MValue {
         match self {
             MValue::String(s) => Ok(s),
             MValue::Timestamp(ts) => Ok(ts.to_string()),
-            MValue::Double(f) | MValue::NullableDouble(Some(f)) => Ok(f.to_string()),
+            MValue::Double(f) => Ok(f.to_string()),
             mv => Err(ConversionError::ValueType(format!(
                 "DbValueInto<String> not implemented for {:?}",
                 mv
@@ -128,7 +128,7 @@ impl DeserializableResultset for Resultset {
         self.number_of_fields()
     }
 
-    fn fieldname(&self, i: usize) -> Option<&String> {
+    fn fieldname(&self, i: usize) -> Option<&str> {
         self.fieldname(i)
     }
 }